+++ tests/test_mimetype_overaly.py
@@ -5,7 +5,7 @@ import os
 from . import tmp_dataset_fixture  # NOQA
 
 
-def test_mimetype_overlay(tmp_dataset_fixture):  # NOQA
+def test_add_mimetype_overlay(tmp_dataset_fixture):  # NOQA
     from dtoolutils.overlays import add_mimetype
 
     # Work around issue with git not storing empty directories.
@@ -41,3 +41,21 @@ def test_mimetype_overlay(tmp_dataset_fixture):  # NOQA
         actual = mimetype_overlay[i]
         expected = expected_mimetypes[fname]
         assert expected == actual
+
+
+def test_add_mimetype_overlay_does_not_raise_when_called_more_than_once(tmp_dataset_fixture):  # NOQA
+    from dtoolutils.overlays import add_mimetype
+
+    # Work around issue with git not storing empty directories.
+    overlays_dir = os.path.join(
+        tmp_dataset_fixture._abs_path,
+        ".dtool",
+        "overlays")
+    if not os.path.isdir(overlays_dir):
+        os.mkdir(overlays_dir)
+
+    assert "mimetype" not in tmp_dataset_fixture.overlays
+
+    add_mimetype(tmp_dataset_fixture)
+    assert "mimetype" in tmp_dataset_fixture.overlays
+    add_mimetype(tmp_dataset_fixture)